refactor(guard): isolate network status validation
